+++ config/releases.exs
@@ -9,7 +9,8 @@ config :td_dd, TdDd.Repo,
   hostname: System.fetch_env!("DB_HOST"),
   port: System.get_env("DB_PORT", "5432") |> String.to_integer(),
   pool_size: System.get_env("DB_POOL_SIZE", "16") |> String.to_integer(),
-  timeout: System.get_env("DB_TIMEOUT_MILLIS", "600000") |> String.to_integer()
+  timeout: System.get_env("DB_TIMEOUT_MILLIS", "600000") |> String.to_integer(),
+  ssl: System.get_env("DB_SSL", "") |> String.downcase() == "true"
 
 config :td_dd, Truedat.Auth.Guardian, secret_key: System.fetch_env!("GUARDIAN_SECRET_KEY")