+++ CHANGELOG.md
@@ -9,8 +9,7 @@
 ### Fixed
 
 - [TD-5670] Structure notes bulk_update when has invalid external_id and template not found
-
-
+- [TD-5696] Rejected implementations cant be edited
 
 ## [5.4.0] 2023-03-28